Geofencing version
I tried to merge the geofencing code with the app, but no matter how hard I tried, I couldn't get it to work, instead, I managed to get everything working in my program, that's why I uploaded a new project. If anyone knows how to make it work on the original that would be great, if not, we can work on this other version.